ci: use pytest (#18)
I skip the only test, since it fails (see
https://github.com/developmentseed/adaptation-atlas-assistant/pull/6/files/3022839fbd2145512c675c434cf569565a3ae23a#r2406028737).
We'll add unit tests, that don't require integration with an external
agent, later.

We might try to set up integration tests on CI, but that might be
trickier b/c we have to hit an external service — we'll see. For now,
the integration tests are in **pytest** b/c that gives us a nice
step-through development workflow.

Closes #16.

---------

Co-authored-by: Emma Paz <ep@emmapaz.com>